requirements.txt
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
safetensors>=0.4.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
tqdm>=4.66.0
open-clip-torch>=2.24.0
datasets>=2.16.0
huggingface-hub>=0.20.0
